refactor: extract WidgetFrame to shared src/components
