+++ src/components/widget-frame.tsx
