+++ src/features/resume/components/widgets/battle-map.tsx
@@ -1,4 +1,4 @@
-import { WidgetFrame } from "./widget-frame";
+import { WidgetFrame } from "@/components/widget-frame";
 
 export function BattleMap() {
   return (
