+++ src/features/resume/components/widgets/compounding-arc.tsx
@@ -1,4 +1,4 @@
-import { WidgetFrame } from "./widget-frame";
+import { WidgetFrame } from "@/components/widget-frame";
 
 export function CompoundingArc() {
   return (
