+++ src/features/resume/components/widgets/reyrey-terminal.tsx
@@ -1,4 +1,4 @@
-import { WidgetFrame } from "./widget-frame";
+import { WidgetFrame } from "@/components/widget-frame";
 
 export function ReyreyTerminal() {
   return (
